server/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-crum==0.7.9
django-extensions==3.2.3
django-jazzmin==3.0.0
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
